requirements.txt
flask
transformers
torch
datasets
scikit-learn
pillow
requests
